requirements.txt
fastapi
uvicorn[standard]
python-docx
pydantic
litellm
python-dotenv
redis
pathspec
tenacity
pytest
fakeredis
pytest-asyncio
httpx
pytest-cov
mammoth
aiofiles
pymysql
PyJWT
